Change `MoveIndex` to be a `u32`
In theory it's possible to have more than 65536 parallel moves due to
block parameters.

+++ src/internal/parallel_moves.rs
@@ -58,8 +58,8 @@ impl EmergencySpillSlotCache {
 
 /// An index into the set of parallel moves that is currently being considered.
 #[derive(Copy, Clone, Eq, PartialEq, Hash, PartialOrd, Ord)]
-struct MoveIndex(u16);
-entity_impl!(MoveIndex(u16), "move");
+struct MoveIndex(u32);
+entity_impl!(MoveIndex(u32), "move");
 
 /// Information about a parallel move.
 struct Move {
